Use a reference for VectorToArray
Because otherwise the vector is copied on the stack (pass by value)
and `a.data()` returns an address in stack memory.

When `VectorToArray()` returns, the data at this address can be
overwritten at any time.

Using a reference avoids the duplication and returns the address
of the original parameter that is still valid in the calling context.